make gonio creation generic
