+++ .vscode/python-artemis-dodal-nexgen.code-workspace
@@ -1,0 +1,19 @@
+{
+	"folders": [
+		{
+			"path": ".."
+		},
+		{
+			"path": "../../dodal"
+		},
+		{
+			"path": "../../nexgen"
+		}
+	],
+	"settings": {
+		"python.languageServer": "Pylance",
+		"terminal.integrated.gpuAcceleration": "off",
+		"esbonio.sphinx.confDir": "",
+		"search.useIgnoreFiles": false
+	}
+}
